requirements.txt
torch
torchvision
PILLOW
numpy
opencv-python